[CHG] multiprocessing: The Jobs workers are affected by the database being closed when they are ran in the Cron Worker process.
Thus, when OpenERP run in multiprocess, the connector do not start job workers threads.
Instead, the new script ``openerp-connector-worker`` should be used. It spawns processes
which start the the job workers threads themselves.

This is not ideal as soon as we have to ensure that OpenERP AND the script are running.
However: it still works normally when OpenERP is not using multiprocessing and this change
allow more control on the worker processes (to implement PG's NOTIFY for instance).