[IMP] refactor the child mappings, using a intermediate class MapChild
This class is responsible for choosing the Mapper class for the items, eventually filter the items, apply the mappings, then format the items (e.g. using the relationship OpenERP commands) and returns the items.

It is not mandatory to explicitly create a MapChild for all the mappings with items (the default one is used).
It can be used for instance to: filter items, change the output format, link items on existing items