Allow configuring runner from config file
In addition to the ODOO_CONNECTOR_CHANNELS environment variable, for
backward compatibility, we can configure the runner channels through
the ODOO_QUEUE_JOB_CHANNELS and in the odoo configuration file like this:

    [options-job_queue]
    channels = root:4

For backward compatibility, the section `[options-connector]` is also
accepted.

Moreover, it's now also possible to preload the module in the config
file, like:

    [options]
    server_wide_modules = web,web_kanban,queue_job

As such we can configure the job runner using only the Odoo config file.