[FIX] queue_job: Complete migration script
As there's a new computed field called `channel_method_name`, if the
computation is done on module upgrade, maybe some models are not
loaded yet, so we pre-create the column for avoiding this calculation
and make it manually at the end of the process. We redo the computing
code relaxing the constraints for avoiding also errors on missing
models or methods.