[IMP] avoid base_sparse_field --load
`base_sparse_field` patches odoo fields
in order to be able to use `fields.Serialized`.

To "pay" this shortcut you have to load it at server load
which makes harder to test it with external tools (ie: pytest)
and disorient first users as they always miss this step.

Now we load the field straight from the module: easy, neat, works.