Add a field to store the overriden channel
This change deserves a commit to explain the reasoning.
In older odoo versions, the code was relying on an implementation detail
of the computed fields mechanisms. Now, the inverse field stores the
value, but the channel field is marked as 'todo' and the computed
function overrides the channel.

This commit adds a new field, used only as storage when we override the
channel. No migration is required because the actual channel is a stored
field.