Support several recipient users in base_export_async
The UI still send it to the current user, but it will allow
reusing the module for other purposes: plan a regular export
to users.

Setting the recipient_ids instead of the email_to in mail.mail means
that we can still send the email, it will stay in the failed emails,
allowing to correct the user.