[12.0][FIX] Force attachment for async import as 'binary'
For unknown reason, possibly a "default_type" key in context, the attachment
has been created with a value "product" for the type, which fails with an
exception because this type is invalid.

Even if the root cause is not known yet, forcing the type as binary is more
defensive and always the value we expect.