[IMP] queue_job: change jobs to paused channel
In some cases, we need to pause only a job since that it is used
to schedule a process for a future date but is needed to filter
only the paused ones.

This commit allows to change a channel job to paused channel
(capacity equals zero) using a wizard.

Fix https://github.com/OCA/queue/issues/166