Update job.py
I had an issue where the execution of the queue job would not except a RetryableJobError, but still end up requeuing itself (because a few rows failed ? honestly not sure). This caused it to never pass in the check for max retries. Putting that check before attempting the job fixed that issue.