[FIX] base_import_async: remove default values from context when creating the attachment.
Some default values might be present in context depending on the action we came from
when clicking on 'import' button. These default values are not intended to
be default values for the ir.attachment record.
In some cases they cause an error because a field with the same name exists on
ir.attachment, as for e.g. the 'default_type'='opportunity' value
present in the standard crm.lead action context.