[FIX] queue_job: indicate job ID in failure message
Failed queue jobs include a button to get to the job when received by email, added automatically by the `mail.message_notification_email` template.

However, when customers forward the email to their support team, they sometimes just copy and paste the error message. Or sometimes they don't have HTML edition available, so as to copy and paste the full mail and include the button.

Thus, it's simpler to just add the job ID to the failure notification. This way, it's easier for the support team to understand the source of the failure.

@moduon MT-3813