[FIX] queue_job: max retry
When job fails because of concurrent update error, it does not respect
max retries set by the job. Problem is ``perform`` method logic that
handles re-try is never called, because ``runjob`` in controller that
triggers jobs, catches expected exception and silences it. Though it
is done to not pollute logs.

So for now, adding extra check before job is run, to make sure max
retries are handled if it reached it.