[bt#28115] queue_job: fix deadlock
Fix a deadlock condition when there are dependent stored fields on `queue.job:state`.

Example:
 - Sale-Order has a stored computed field which depends on job-state.
 - A job updates a sale-order record.
 - A exception is raised
and in the final error-handling with a child cursor in https://github.com/OCA/queue/blob/c93a25366d2c51339ae4530a42eb7393bdc98b12/queue_job/controllers/main.py#L144
the job's state is updated, which also triggers an update in the sale order.
Result is a deadlock.

This commit adds a rollback in case of error during job execution, so that any locks are released.